Prevent interruptions caused by UNKNOWN_TOPIC_OR_PARTITION
When a client commits a transaction the commit method waits until the
pending publish futures are resolved. However if a future throws
UnknownTopicOrPartitionException (which is supposed to be retryable)
the commit methods fails too and the only thing to do is to retire
current session and to create a new producer.

We want to avoid interruptions and creating a new producer so we use
a workaround: we explicitly wait until the futures are resolved before
invoking commit. In case a future fails we invoke abort instead of
commit and continue using the same producer.